refactor(core): decompose /percy/maestro-screenshot relay (incl. self-hosted) into modules
Extract the ~517-line maestro-screenshot handler out of api.js into cohesive
modules, folding in #2261's self-hosted branches:
- maestro-screenshot.js — handleMaestroScreenshot orchestrator + parsePngDimensions;
  runtime detection (selfHosted = !sessionId), conditional sessionId validation,
  PERCY_MAESTRO_SCREENSHOT_DIR scope-root resolution + filePath rejection.
- maestro-screenshot-file.js — locateScreenshot({...,scopeRoot,selfHosted}):
  BS session glob + self-hosted recursive dot:true glob (Windows-normalized) +
  manualScreenshotWalk (BS only) + realpath/scopeRoot prefix check.
- maestro-regions.js — validateRegionInputs + resolveRegions (explicit-key merge).
api.js is now a route table (1072 → 382 lines) delegating via named imports;
dead ServerError/maestro-hierarchy imports dropped. semgrep path-traversal
suppression retargeted to maestro-screenshot-file.js (+ api.js createStaticServer).
Behavior-preserving: api.test.js (BS + self-hosted) passes unchanged.
