+++ packages/core/src/maestro-screenshot.js
@@ -1,0 +1,231 @@
+import fs from 'fs';
+import path from 'path';
+import { normalize } from '@percy/config/utils';
+import { ServerError } from './server.js';
+import { encodeURLSearchParams } from './utils.js';
+import { handleSyncJob } from './snapshot.js';
+import { locateScreenshot } from './maestro-screenshot-file.js';
+import { validateRegionInputs, resolveRegions } from './maestro-regions.js';
+
+// Parse PNG IHDR chunk for the screenshot's actual rendered dimensions.
+// Returns { width, height } when the buffer is a valid PNG with non-zero
+// dimensions, or null otherwise (non-PNG signature, truncated file, zero
+// IHDR values). PNG layout per W3C spec:
+//   bytes 0..7   PNG signature (89 50 4E 47 0D 0A 1A 0A)
+//   bytes 8..15  IHDR chunk header (length + type, fixed)
+//   bytes 16..19 width  (big-endian uint32)
+//   bytes 20..23 height (big-endian uint32)
+// No library dependency — pure stdlib Buffer access on the bytes the relay
+// has already read.
+export function parsePngDimensions(buffer) {
+  if (!buffer || buffer.length < 24) return null;
+  if (buffer[0] !== 0x89 || buffer[1] !== 0x50 || buffer[2] !== 0x4E || buffer[3] !== 0x47 ||
+      buffer[4] !== 0x0D || buffer[5] !== 0x0A || buffer[6] !== 0x1A || buffer[7] !== 0x0A) {
+    return null;
+  }
+  const width = buffer.readUInt32BE(16);
+  const height = buffer.readUInt32BE(20);
+  if (width <= 0 || height <= 0) return null;
+  return { width, height };
+}
+
+// Handler for `post /percy/maestro-screenshot`: post a comparison by reading
+// a Maestro screenshot from disk.
+export async function handleMaestroScreenshot(req, res, percy) {
+  /* istanbul ignore next — req.body falsy guard; tests always pass a body. */
+  let { name, sessionId } = req.body || {};
+
+  if (!name) throw new ServerError(400, 'Missing required field: name');
+
+  // Self-hosted vs BrowserStack is signaled by sessionId presence: BS
+  // host-injection always supplies it; self-hosted runs never do.
+  let selfHosted = !sessionId;
+
+  // Strict character-class validation — rejects path separators, shell metacharacters,
+  // NUL, newlines, and anything else that could confuse the glob or the filesystem.
+  // `name` is load-bearing for the recursive glob — must not be loosened.
+  const SAFE_ID = /^[a-zA-Z0-9_-]+$/;
+  if (!SAFE_ID.test(name)) {
+    throw new ServerError(400, 'Invalid screenshot name');
+  }
+  if (sessionId && !SAFE_ID.test(sessionId)) {
+    throw new ServerError(400, 'Invalid sessionId');
+  }
+
+  // Resolve platform signal: strict whitelist on `platform` when present; default Android when absent.
+  // Backward compatible with SDK v0.2.0 (no platform field → Android glob).
+  let platform = 'android';
+  if (req.body.platform !== undefined) {
+    if (typeof req.body.platform !== 'string') {
+      throw new ServerError(400, 'Invalid platform: must be a string');
+    }
+    let normalized = req.body.platform.toLowerCase();
+    if (normalized !== 'ios' && normalized !== 'android') {
+      throw new ServerError(400, `Invalid platform: must be "ios" or "android", got "${req.body.platform}"`);
+    }
+    platform = normalized;
+  }
+
+  // Optional caller-supplied absolute path. When present, the relay reads
+  // the file directly and skips the legacy glob — the SDK has already
+  // chosen the path under the BS session root. Shape errors (non-string,
+  // non-absolute, too long) are 400. Existence and session-root scoping
+  // are enforced by the shared realpath + prefix check below, which
+  // returns 404 — same shape as the glob path. Treat empty string as
+  // absent so older SDKs that emit the field unconditionally still fall
+  // through to the glob.
+  let suppliedFilePath = null;
+  if (req.body.filePath !== undefined && req.body.filePath !== null && req.body.filePath !== '') {
+    if (typeof req.body.filePath !== 'string') {
+      throw new ServerError(400, 'Invalid filePath: must be a string');
+    }
+    if (req.body.filePath.length > 1024) {
+      throw new ServerError(400, 'Invalid filePath: exceeds maximum length of 1024');
+    }
+    if (!path.isAbsolute(req.body.filePath)) {
+      throw new ServerError(400, 'Invalid filePath: must be an absolute path');
+    }
+    suppliedFilePath = req.body.filePath;
+  }
+
+  // Resolve the file-find scope root. On BrowserStack (sessionId present), the
+  // root is the BS host's /tmp/{sessionId}{_test_suite} convention. Self-hosted
+  // (sessionId absent) requires PERCY_MAESTRO_SCREENSHOT_DIR (read from
+  // process.env, never the request body) to be an absolute, existing directory
+  // — typically the customer's `maestro test --test-output-dir <DIR>` path. The
+  // realpath + prefix check inside locateScreenshot enforces the security
+  // invariant at whichever root applies; the boundary is relocated, not removed.
+  let scopeRoot;
+  if (selfHosted) {
+    // Reject filePath outright in self-hosted mode. The SDK never emits it (it
+    // sends a relative SCREENSHOT_NAME); honoring an absolute filePath against
+    // a caller-influenceable root would re-open arbitrary in-root reads.
+    if (suppliedFilePath) {
+      throw new ServerError(400, 'filePath is not accepted in self-hosted mode (omit it; PERCY_MAESTRO_SCREENSHOT_DIR + relative SCREENSHOT_NAME is the supported path)');
+    }
+    let dir = process.env.PERCY_MAESTRO_SCREENSHOT_DIR;
+    if (!dir) {
+      throw new ServerError(400, 'Missing required env: PERCY_MAESTRO_SCREENSHOT_DIR (set it to your `maestro test --test-output-dir` path)');
+    }
+    if (!path.isAbsolute(dir)) {
+      throw new ServerError(400, 'PERCY_MAESTRO_SCREENSHOT_DIR must be an absolute path');
+    }
+    // UX guard ONLY: surface an actionable 400 ("dir not found") instead of the
+    // opaque 404 the realpath+prefix containment check would emit for a missing
+    // dir. A small TOCTOU window exists between this stat and locateScreenshot's
+    // realpath — acceptable because realpath (not stat) is the security
+    // invariant: even if the dir is swapped for a symlink in between, realpath
+    // resolves the target and the sep-prefix check rejects anything outside it.
+    let stat;
+    try { stat = await fs.promises.stat(dir); } catch { stat = null; }
+    if (!stat || !stat.isDirectory()) {
+      throw new ServerError(400, `PERCY_MAESTRO_SCREENSHOT_DIR is not an existing directory: ${dir}`);
+    }
+    scopeRoot = dir;
+  } else {
+    scopeRoot = platform === 'ios'
+      ? `/tmp/${sessionId}`
+      : `/tmp/${sessionId}_test_suite`;
+  }
+
+  // Validate regions input shape early (before file I/O and ADB work) so
+  // malformed requests don't consume resolver/relay work.
+  validateRegionInputs(req.body);
+
+  // Locate the screenshot on disk (supplied filePath, BS session glob, or
+  // self-hosted PERCY_MAESTRO_SCREENSHOT_DIR recursive glob) and confirm it
+  // resolves under scopeRoot. Throws ServerError(404) when missing/out-of-root.
+  let realPath = await locateScreenshot({ platform, sessionId, name, filePath: suppliedFilePath, scopeRoot, selfHosted });
+
+  // Read and base64-encode the screenshot
+  let fileContent = await fs.promises.readFile(realPath);
+  let base64Content = fileContent.toString('base64');
+
+  // Parse the PNG header for actual rendered dimensions. The PNG bytes
+  // ARE the source of truth — what Percy stores and compares against.
+  // Fills tag.width/height when the customer didn't supply them (or
+  // supplied invalid values); customer-supplied values continue to win
+  // for backward compat with any flow that pins a specific tag dim.
+  let pngDims = parsePngDimensions(fileContent);
+
+  // Build tag from optional request body fields
+  let tag = req.body.tag || { name: 'Unknown Device', osName: 'Android' };
+  /* istanbul ignore if — fallback when tag.name is missing; tests always
+     pass a complete tag object. */
+  if (!tag.name) tag.name = 'Unknown Device';
+  if (pngDims) {
+    if (typeof tag.width !== 'number' || tag.width <= 0 || isNaN(tag.width)) {
+      tag.width = pngDims.width;
+    }
+    if (typeof tag.height !== 'number' || tag.height <= 0 || isNaN(tag.height)) {
+      tag.height = pngDims.height;
+    }
+  }
+
+  // Construct comparison payload with tile metadata from request
+  let payload = {
+    name,
+    tag,
+    tiles: [{
+      content: base64Content,
+      statusBarHeight: req.body.statusBarHeight || 0,
+      navBarHeight: req.body.navBarHeight || 0,
+      headerHeight: 0,
+      footerHeight: 0,
+      fullscreen: req.body.fullscreen || false
+    }],
+    clientInfo: req.body.clientInfo || 'percy-maestro/0.1.0',
+    environmentInfo: req.body.environmentInfo || 'percy-maestro'
+  };
+
+  if (req.body.testCase) payload.testCase = req.body.testCase;
+  if (req.body.labels) payload.labels = req.body.labels;
+  if (req.body.thTestCaseExecutionId) payload.thTestCaseExecutionId = req.body.thTestCaseExecutionId;
+
+  // Resolve element/coordinate regions to comparison-payload fragments.
+  // Element regions degrade gracefully — a resolver failure warn-skips those
+  // regions only; the snapshot still uploads. The hierarchy dump is memoized
+  // per request inside resolveRegions. Assign the three known comparison keys
+  // explicitly (never Object.assign of request-derived data) so only these
+  // fields can ever be set here.
+  let regionData = await resolveRegions({ body: req.body, platform, sessionId, percy });
+  if (regionData.regions) payload.regions = regionData.regions;
+  if (regionData.ignoredElementsData) payload.ignoredElementsData = regionData.ignoredElementsData;
+  if (regionData.consideredElementsData) payload.consideredElementsData = regionData.consideredElementsData;
+
+  // Upload via percy — sync or fire-and-forget
+  if (req.body.sync === true) payload.sync = true;
+
+  let data;
+  if (percy.syncMode(payload)) {
+    // percy.upload returns an async generator that must be drained for #snapshots.push to run.
+    // See docs/solutions/best-practices/2026-05-20-maestro-sync-promise-bug-investigation.md.
+    const snapshotPromise = new Promise((resolve, reject) => {
+      const upload = percy.upload(payload, { resolve, reject }, 'app');
+      (async () => {
+        // eslint-disable-next-line no-unused-vars
+        try { for await (const _ of upload) { /* drain */ } } catch (e) { reject(e); }
+      })();
+    });
+    data = await handleSyncJob(snapshotPromise, percy, 'comparison');
+    return res.json(200, { success: true, data });
+  }
+
+  let upload = percy.upload(payload, null, 'app');
+  /* istanbul ignore if — ?await=true URL flag triggers fire-and-wait;
+     tests cover both syncMode and fire-and-forget but not the explicit
+     ?await query-param variant. */
+  if (req.url.searchParams.has('await')) await upload;
+
+  // Generate redirect link
+  let link = [
+    percy.client.apiUrl, '/comparisons/redirect?',
+    encodeURLSearchParams(normalize({
+      buildId: percy.build?.id,
+      snapshot: { name },
+      tag
+    }, { snake: true }))
+  ].join('');
+
+  return res.json(200, { success: true, link });
+}
